# Optimize tests to run in parallel