--- a/common/barrier_test.go
+++ b/common/barrier_test.go
@@ -11,6 +11,8 @@ import (
 )
 
 func TestBarrier(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 
 	log := log.NewNullLogger()