--- a/common/browser_context_test.go
+++ b/common/browser_context_test.go
@@ -14,7 +14,11 @@ import (
 )
 
 func TestNewBrowserContext(t *testing.T) {
+	t.Parallel()
+
 	t.Run("add_web_vital_js_scripts_to_context", func(t *testing.T) {
+		t.Parallel()
+
 		ctx, cancel := context.WithCancel(context.Background())
 		logger := log.NewNullLogger()
 		b := newBrowser(ctx, cancel, nil, NewLocalBrowserOptions(), logger)