--- a/tests/browser_context_options_test.go
+++ b/tests/browser_context_options_test.go
@@ -38,6 +38,8 @@ func TestBrowserContextOptionsDefaultValues(t *testing.T) {
 }
 
 func TestBrowserContextOptionsDefaultViewport(t *testing.T) {
+	t.Parallel()
+
 	p := newTestBrowser(t).NewPage(nil)
 
 	viewportSize := p.ViewportSize()
@@ -46,6 +48,8 @@ func TestBrowserContextOptionsDefaultViewport(t *testing.T) {
 }
 
 func TestBrowserContextOptionsSetViewport(t *testing.T) {
+	t.Parallel()
+
 	tb := newTestBrowser(t)
 	bctx, err := tb.NewContext(tb.toGojaValue(struct {
 		Viewport common.Viewport `js:"viewport"`
@@ -66,6 +70,8 @@ func TestBrowserContextOptionsSetViewport(t *testing.T) {
 }
 
 func TestBrowserContextOptionsExtraHTTPHeaders(t *testing.T) {
+	t.Parallel()
+
 	tb := newTestBrowser(t, withHTTPServer())
 	bctx, err := tb.NewContext(tb.toGojaValue(struct {
 		ExtraHTTPHeaders map[string]string `js:"extraHTTPHeaders"`