--- a/tests/js_handle_get_properties_test.go
+++ b/tests/js_handle_get_properties_test.go
@@ -9,6 +9,8 @@ import (
 )
 
 func TestJSHandleGetProperties(t *testing.T) {
+	t.Parallel()
+
 	tb := newTestBrowser(t)
 	p := tb.NewPage(nil)