--- a/tests/keyboard_test.go
+++ b/tests/keyboard_test.go
@@ -12,7 +12,11 @@ import (
 )
 
 func TestKeyboardPress(t *testing.T) {
+	t.Parallel()
+
 	t.Run("all_keys", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()
@@ -26,6 +30,8 @@ func TestKeyboardPress(t *testing.T) {
 	})
 
 	t.Run("backspace", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()
@@ -43,6 +49,8 @@ func TestKeyboardPress(t *testing.T) {
 	})
 
 	t.Run("combo", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()
@@ -68,6 +76,7 @@ func TestKeyboardPress(t *testing.T) {
 	})
 
 	t.Run("meta", func(t *testing.T) {
+		t.Parallel()
 		t.Skip("FIXME") // See https://github.com/grafana/xk6-browser/issues/424
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
@@ -94,6 +103,8 @@ func TestKeyboardPress(t *testing.T) {
 	})
 
 	t.Run("type does not split on +", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()
@@ -108,6 +119,8 @@ func TestKeyboardPress(t *testing.T) {
 	})
 
 	t.Run("capitalization", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()
@@ -134,6 +147,8 @@ func TestKeyboardPress(t *testing.T) {
 	})
 
 	t.Run("type not affected by shift", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()
@@ -151,6 +166,8 @@ func TestKeyboardPress(t *testing.T) {
 	})
 
 	t.Run("newline", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()
@@ -169,6 +186,8 @@ func TestKeyboardPress(t *testing.T) {
 
 	// Replicates the test from https://playwright.dev/docs/api/class-keyboard
 	t.Run("selection", func(t *testing.T) {
+		t.Parallel()
+
 		tb := newTestBrowser(t)
 		p := tb.NewPage(nil)
 		kb := p.GetKeyboard()