--- a/tests/locator_test.go
+++ b/tests/locator_test.go
@@ -434,6 +434,8 @@ func TestLocatorElementState(t *testing.T) {
 }
 
 func TestLocatorPress(t *testing.T) {
+	t.Parallel()
+
 	tb := newTestBrowser(t)
 
 	p := tb.NewPage(nil)