## What?

1. Makes unit and integration tests run in parallel.
2. Makes some CI optimizations, considering that this is an optimization PR. Let me know if that part deserves another PR.

## Why?

1. The current tests take about 1 minute to run (locally), and there is no reason to run tests sequentially.
2. This also lets us detect race conditions much faster and prevent unintended deadlocks in code.
3. The E2E runner is hard to debug when running without the `-q` flag. With the flag, the result looks clearer to read and debug without the k6 banner and progress bar.

Here are the results after this change (On a 10-core machine):

```
Before: ~48s
After : ~24s

===Before===
ok      github.com/grafana/xk6-browser/browser  1.853s
ok      github.com/grafana/xk6-browser/common   1.669s
ok      github.com/grafana/xk6-browser/chromium 1.511s
ok      github.com/grafana/xk6-browser/tests    46.902s

===After===
ok      github.com/grafana/xk6-browser/browser  0.956s
ok      github.com/grafana/xk6-browser/common   0.689s
ok      github.com/grafana/xk6-browser/chromium 0.251s
ok      github.com/grafana/xk6-browser/tests    23.150s
```

On CI, it's the same because our runner runs on a 2-core. We can make this much faster if/when we run it on a more powerful runner. 

## Checklist

- [x] I have performed a self-review of my code
- [x] I have added tests for my changes
- [x] I have commented on my code, particularly in hard-to-understand areas

## Related PR(s)/Issue(s)

Updates #1018