@inancgumus 2023-08-31
Thank you for the review 🙇

> Also, I think there are a few tests missing to be executed in parallel or add a comment explaining why the can't not be run as such.

> TestBrowserContextOptionsPermissions, TestConnection, TestConnectionClosureAbnormal, TestConnectionSendRecv, TestConnectionCreateSession, TestContextWithDoneChan, TestPageLocator, TestLocatorPress.

Ah, I missed these ones; thanks! I'll make them parallel.

> TestNewBrowserContext

I had sent this PR before our today's discussion in #1007. I was expecting #1007 to use globals, and it wouldn't make sense for this test to be parallel. After the discussion we had, we can make it parallel. WDYT?